feat(step1): 상품의 가격을 변경할 수 있다.

+++ src/test/java/kitchenpos/products/tobe/infrastructure/JdbcTemplateProductRepository.java
@@ -78,7 +78,7 @@ public class JdbcTemplateProductRepository implements ProductRepository {
                 resultSet.getLong(KEY_COLUMN),
                 resultSet.getString(NAME_COLUMN),
                 new ProfanityName(),
-                resultSet.getBigDecimal(PRICE_COLUMN)
+                resultSet.getLong(PRICE_COLUMN)
         );
     }
 }
